% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{BasicPerfSampleSeries}
\alias{BasicPerfSampleSeries}
\title{BasicPerfSampleSeries Object}
\usage{
BasicPerfSampleSeries(perfMetricType = NULL, perfUnit = NULL,
  sampleSeriesLabel = NULL)
}
\arguments{
\item{perfMetricType}{No description}

\item{perfUnit}{No description}

\item{sampleSeriesLabel}{No description}
}
\value{
BasicPerfSampleSeries object
}
\description{
BasicPerfSampleSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulates the metadata for basic sample series represented by a line chart
}
