% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{BatchCreatePerfSamplesRequest}
\alias{BatchCreatePerfSamplesRequest}
\title{BatchCreatePerfSamplesRequest Object}
\usage{
BatchCreatePerfSamplesRequest(perfSamples = NULL)
}
\arguments{
\item{perfSamples}{The set of PerfSamples to create should not include existing timestamps}
}
\value{
BatchCreatePerfSamplesRequest object
}
\description{
BatchCreatePerfSamplesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request must provide up to a maximum of 5000 samples to be created; a larger sample size will cause an INVALID_ARGUMENT error
}
