% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{CPUInfo}
\alias{CPUInfo}
\title{CPUInfo Object}
\usage{
CPUInfo(cpuProcessor = NULL, cpuSpeedInGhz = NULL, numberOfCores = NULL)
}
\arguments{
\item{cpuProcessor}{description of the device processor ie '1}

\item{cpuSpeedInGhz}{the CPU clock speed in GHz}

\item{numberOfCores}{the number of CPU cores}
}
\value{
CPUInfo object
}
\description{
CPUInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
