% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Execution}
\alias{Execution}
\title{Execution Object}
\usage{
Execution(completionTime = NULL, creationTime = NULL, executionId = NULL,
  outcome = NULL, state = NULL, testExecutionMatrixId = NULL)
}
\arguments{
\item{completionTime}{The time when the Execution status transitioned to COMPLETE}

\item{creationTime}{The time when the Execution was created}

\item{executionId}{A unique identifier within a History for this Execution}

\item{outcome}{Classify the result, for example into SUCCESS or FAILURE}

\item{state}{The initial state is IN_PROGRESS}

\item{testExecutionMatrixId}{TestExecution Matrix ID that the Test Service uses}
}
\value{
Execution object
}
\description{
Execution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Execution represents a collection of Steps. For instance, it could represent: - a mobile test executed across a range of device configurations - a jenkins job with a build step followed by a test stepThe maximum size of an execution message is 1 MiB.An Execution can be updated until its state is set to COMPLETE at which point it becomes immutable.
}
