% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{FailureDetail}
\alias{FailureDetail}
\title{FailureDetail Object}
\usage{
FailureDetail(crashed = NULL, notInstalled = NULL,
  otherNativeCrash = NULL, timedOut = NULL, unableToCrawl = NULL)
}
\arguments{
\item{crashed}{If the failure was severe because the system under test crashed}

\item{notInstalled}{If an app is not installed and thus no test can be run with the app}

\item{otherNativeCrash}{If a native process other than the app crashed}

\item{timedOut}{If the test overran some time limit, and that is why it failed}

\item{unableToCrawl}{If the robo was unable to crawl the app; perhaps because the app did not start}
}
\value{
FailureDetail object
}
\description{
FailureDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
