% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{FileReference}
\alias{FileReference}
\title{FileReference Object}
\usage{
FileReference(fileUri = NULL)
}
\arguments{
\item{fileUri}{The URI of a file stored in Google Cloud Storage}
}
\value{
FileReference object
}
\description{
FileReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a file.
}
