% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{History}
\alias{History}
\title{History Object}
\usage{
History(displayName = NULL, historyId = NULL, name = NULL)
}
\arguments{
\item{displayName}{A short human-readable (plain text) name to display in the UI}

\item{historyId}{A unique identifier within a project for this History}

\item{name}{A name to uniquely identify a history within a project}
}
\value{
History object
}
\description{
History Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A History represents a sorted list of Executions ordered by the start_timestamp_millis field (descending). It can be used to group all the Executions of a continuous build.Note that the ordering only operates on one-dimension. If a repository has multiple branches, it means that multiple histories will need to be used in order to order Executions per branch.
}
