% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(error = NULL, sourceImage = NULL, stepId = NULL, thumbnail = NULL)
}
\arguments{
\item{error}{An error explaining why the thumbnail could not be rendered}

\item{sourceImage}{A reference to the full-size, original image}

\item{stepId}{The step to which the image is attached}

\item{thumbnail}{The thumbnail}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An image, with a link to the main image and a thumbnail.
}
