% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{InconclusiveDetail}
\alias{InconclusiveDetail}
\title{InconclusiveDetail Object}
\usage{
InconclusiveDetail(abortedByUser = NULL, infrastructureFailure = NULL)
}
\arguments{
\item{abortedByUser}{If the end user aborted the test execution before a pass or fail could be determined}

\item{infrastructureFailure}{If the test runner could not determine success or failure because the test depends on a component other than the system under test which failed}
}
\value{
InconclusiveDetail object
}
\description{
InconclusiveDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
