% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListExecutionsResponse}
\alias{ListExecutionsResponse}
\title{ListExecutionsResponse Object}
\usage{
ListExecutionsResponse(executions = NULL, nextPageToken = NULL)
}
\arguments{
\item{executions}{Executions}

\item{nextPageToken}{A continuation token to resume the query at the next item}
}
\value{
ListExecutionsResponse object
}
\description{
ListExecutionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
