% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListHistoriesResponse}
\alias{ListHistoriesResponse}
\title{ListHistoriesResponse Object}
\usage{
ListHistoriesResponse(histories = NULL, nextPageToken = NULL)
}
\arguments{
\item{histories}{Histories}

\item{nextPageToken}{A continuation token to resume the query at the next item}
}
\value{
ListHistoriesResponse object
}
\description{
ListHistoriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for HistoryService.List
}
