% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListPerfSamplesResponse}
\alias{ListPerfSamplesResponse}
\title{ListPerfSamplesResponse Object}
\usage{
ListPerfSamplesResponse(nextPageToken = NULL, perfSamples = NULL)
}
\arguments{
\item{nextPageToken}{Optional, returned if result size exceeds the page size specified in the request (or the default page size, 500, if unspecified)}

\item{perfSamples}{No description}
}
\value{
ListPerfSamplesResponse object
}
\description{
ListPerfSamplesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
