% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListStepsResponse}
\alias{ListStepsResponse}
\title{ListStepsResponse Object}
\usage{
ListStepsResponse(nextPageToken = NULL, steps = NULL)
}
\arguments{
\item{nextPageToken}{A continuation token to resume the query at the next item}

\item{steps}{Steps}
}
\value{
ListStepsResponse object
}
\description{
ListStepsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for StepService.List.
}
