% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ListStepThumbnailsResponse}
\alias{ListStepThumbnailsResponse}
\title{ListStepThumbnailsResponse Object}
\usage{
ListStepThumbnailsResponse(nextPageToken = NULL, thumbnails = NULL)
}
\arguments{
\item{nextPageToken}{A continuation token to resume the query at the next item}

\item{thumbnails}{A list of image data}
}
\value{
ListStepThumbnailsResponse object
}
\description{
ListStepThumbnailsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response containing the thumbnails in a step.
}
