% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{PerfEnvironment}
\alias{PerfEnvironment}
\title{PerfEnvironment Object}
\usage{
PerfEnvironment(cpuInfo = NULL, memoryInfo = NULL)
}
\arguments{
\item{cpuInfo}{CPU related environment info}

\item{memoryInfo}{Memory related environment info}
}
\value{
PerfEnvironment object
}
\description{
PerfEnvironment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulates performance environment info
}
