% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{PerfSample}
\alias{PerfSample}
\title{PerfSample Object}
\usage{
PerfSample(sampleTime = NULL, value = NULL)
}
\arguments{
\item{sampleTime}{Timestamp of collection}

\item{value}{Value observed}
}
\value{
PerfSample object
}
\description{
PerfSample Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Resource representing a single performance measure or data point
}
