% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{PerfSampleSeries}
\alias{PerfSampleSeries}
\title{PerfSampleSeries Object}
\usage{
PerfSampleSeries(basicPerfSampleSeries = NULL, executionId = NULL,
  historyId = NULL, projectId = NULL, sampleSeriesId = NULL,
  stepId = NULL)
}
\arguments{
\item{basicPerfSampleSeries}{Basic series represented by a line chart}

\item{executionId}{A tool results execution ID}

\item{historyId}{A tool results history ID}

\item{projectId}{The cloud project}

\item{sampleSeriesId}{A sample series id}

\item{stepId}{A tool results step ID}
}
\value{
PerfSampleSeries object
}
\description{
PerfSampleSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Resource representing a collection of performance samples (or data points)
}
