% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_functions.R
\name{projects.getSettings}
\alias{projects.getSettings}
\title{Gets the Tool Results settings for a project.May return any of the following canonical error codes:- PERMISSION_DENIED - if the user is not authorized to read from project}
\usage{
projects.getSettings(projectId)
}
\arguments{
\item{projectId}{A Project id}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://firebase.google.com/docs/test-lab/}{Google Documentation}
}
