% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_functions.R
\name{projects.initializeSettings}
\alias{projects.initializeSettings}
\title{Creates resources for settings which have not yet been set.Currently, this creates a single resource: a Google Cloud Storage bucket, to be used as the default bucket for this project. The bucket is created in the name of the user calling. Except in rare cases, calling this method in parallel from multiple clients will only create a single bucket. In order to avoid unnecessary storage charges, the bucket is configured to automatically delete objects older than 90 days.The bucket is created with the project-private ACL: All project team members are given permissions to the bucket and objects created within it according to their roles. Project owners have owners rights, and so on. The default ACL on objects created in the bucket is project-private as well. See Google Cloud Storage documentation for more details.If there is already a default bucket set and the project can access the bucket, this call does nothing. However, if the project doesn't have the permission to access the bucket or the bucket is deteleted, a new bucket will be created.May return any canonical error codes, including the following:- PERMISSION_DENIED - if the user is not authorized to write to project - Any error code raised by Google Cloud Storage}
\usage{
projects.initializeSettings(projectId)
}
\arguments{
\item{projectId}{A Project id}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://firebase.google.com/docs/test-lab/}{Google Documentation}
}
