% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ProjectSettings}
\alias{ProjectSettings}
\title{ProjectSettings Object}
\usage{
ProjectSettings(defaultBucket = NULL, name = NULL)
}
\arguments{
\item{defaultBucket}{The name of the Google Cloud Storage bucket to which results are written}

\item{name}{The name of the project's settings}
}
\value{
ProjectSettings object
}
\description{
ProjectSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Per-project settings for the Tool Results service.
}
