% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{PublishXunitXmlFilesRequest}
\alias{PublishXunitXmlFilesRequest}
\title{PublishXunitXmlFilesRequest Object}
\usage{
PublishXunitXmlFilesRequest(xunitXmlFiles = NULL)
}
\arguments{
\item{xunitXmlFiles}{URI of the Xunit XML files to publish}
}
\value{
PublishXunitXmlFilesRequest object
}
\description{
PublishXunitXmlFilesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for StepService.PublishXunitXmlFiles.
}
