% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{SkippedDetail}
\alias{SkippedDetail}
\title{SkippedDetail Object}
\usage{
SkippedDetail(incompatibleAppVersion = NULL,
  incompatibleArchitecture = NULL, incompatibleDevice = NULL)
}
\arguments{
\item{incompatibleAppVersion}{If the App doesn't support the specific API level}

\item{incompatibleArchitecture}{If the App doesn't run on the specific architecture, for example, x86}

\item{incompatibleDevice}{If the requested OS version doesn't run on the specific device model}
}
\value{
SkippedDetail object
}
\description{
SkippedDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
