% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{StackTrace}
\alias{StackTrace}
\title{StackTrace Object}
\usage{
StackTrace(exception = NULL)
}
\arguments{
\item{exception}{The stack trace message}
}
\value{
StackTrace object
}
\description{
StackTrace Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A stacktrace.
}
