% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{StepDimensionValueEntry}
\alias{StepDimensionValueEntry}
\title{StepDimensionValueEntry Object}
\usage{
StepDimensionValueEntry(key = NULL, value = NULL)
}
\arguments{
\item{key}{No description}

\item{value}{No description}
}
\value{
StepDimensionValueEntry object
}
\description{
StepDimensionValueEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
