% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{SuccessDetail}
\alias{SuccessDetail}
\title{SuccessDetail Object}
\usage{
SuccessDetail(otherNativeCrash = NULL)
}
\arguments{
\item{otherNativeCrash}{If a native process other than the app crashed}
}
\value{
SuccessDetail object
}
\description{
SuccessDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
