% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestCaseReference}
\alias{TestCaseReference}
\title{TestCaseReference Object}
\usage{
TestCaseReference(className = NULL, name = NULL, testSuiteName = NULL)
}
\arguments{
\item{className}{The name of the class}

\item{name}{The name of the test case}

\item{testSuiteName}{The name of the test suite to which this test case belongs}
}
\value{
TestCaseReference object
}
\description{
TestCaseReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a test case.Test case references are canonically ordered lexicographically by these three factors: * First, by test_suite_name. * Second, by class_name. * Third, by name.
}
