% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestExecutionStep}
\alias{TestExecutionStep}
\title{TestExecutionStep Object}
\usage{
TestExecutionStep(testIssues = NULL, testSuiteOverviews = NULL,
  testTiming = NULL, toolExecution = NULL)
}
\arguments{
\item{testIssues}{Issues observed during the test execution}

\item{testSuiteOverviews}{List of test suite overview contents}

\item{testTiming}{The timing break down of the test execution}

\item{toolExecution}{Represents the execution of the test runner}
}
\value{
TestExecutionStep object
}
\description{
TestExecutionStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A step that represents running tests.It accepts ant-junit xml files which will be parsed into structured test results by the service. Xml file paths are updated in order to append more files, however they can't be deleted.Users can also add test results manually by using the test_result field.
}
