% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestIssue}
\alias{TestIssue}
\title{TestIssue Object}
\usage{
TestIssue(errorMessage = NULL, stackTrace = NULL)
}
\arguments{
\item{errorMessage}{A brief human-readable message describing the abnormal event}

\item{stackTrace}{Optional}
}
\value{
TestIssue object
}
\description{
TestIssue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An abnormal event observed during the test execution.
}
