% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestSuiteOverview}
\alias{TestSuiteOverview}
\title{TestSuiteOverview Object}
\usage{
TestSuiteOverview(errorCount = NULL, failureCount = NULL, name = NULL,
  skippedCount = NULL, totalCount = NULL, xmlSource = NULL)
}
\arguments{
\item{errorCount}{Number of test cases in error, typically set by the service by parsing the xml_source}

\item{failureCount}{Number of failed test cases, typically set by the service by parsing the xml_source}

\item{name}{The name of the test suite}

\item{skippedCount}{Number of test cases not run, typically set by the service by parsing the xml_source}

\item{totalCount}{Number of test cases, typically set by the service by parsing the xml_source}

\item{xmlSource}{If this test suite was parsed from XML, this is the URI where the original XML file is stored}
}
\value{
TestSuiteOverview object
}
\description{
TestSuiteOverview Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A summary of a test suite result either parsed from XML or uploaded directly by a user.Note: the API related comments are for StepService only. This message is also being used in ExecutionService in a read only mode for the corresponding step.
}
