% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{TestTiming}
\alias{TestTiming}
\title{TestTiming Object}
\usage{
TestTiming(testProcessDuration = NULL)
}
\arguments{
\item{testProcessDuration}{How long it took to run the test process}
}
\value{
TestTiming object
}
\description{
TestTiming Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Testing timing break down to know phases.
}
