% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{Thumbnail}
\alias{Thumbnail}
\title{Thumbnail Object}
\usage{
Thumbnail(contentType = NULL, data = NULL, heightPx = NULL,
  widthPx = NULL)
}
\arguments{
\item{contentType}{The thumbnail's content type, i}

\item{data}{The thumbnail file itself}

\item{heightPx}{The height of the thumbnail, in pixels}

\item{widthPx}{The width of the thumbnail, in pixels}
}
\value{
Thumbnail object
}
\description{
Thumbnail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single thumbnail, with its size and format.
}
