% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ToolExecution}
\alias{ToolExecution}
\title{ToolExecution Object}
\usage{
ToolExecution(commandLineArguments = NULL, exitCode = NULL,
  toolLogs = NULL, toolOutputs = NULL)
}
\arguments{
\item{commandLineArguments}{The full tokenized command line including the program name (equivalent to argv in a C program)}

\item{exitCode}{Tool execution exit code}

\item{toolLogs}{References to any plain text logs output the tool execution}

\item{toolOutputs}{References to opaque files of any format output by the tool execution}
}
\value{
ToolExecution object
}
\description{
ToolExecution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An execution of an arbitrary tool. It could be a test runner or a tool copying artifacts or deploying code.
}
