% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ToolExecutionStep}
\alias{ToolExecutionStep}
\title{ToolExecutionStep Object}
\usage{
ToolExecutionStep(toolExecution = NULL)
}
\arguments{
\item{toolExecution}{A Tool execution}
}
\value{
ToolExecutionStep object
}
\description{
ToolExecutionStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Generic tool step to be used for binaries we do not explicitly support. For example: running cp to copy artifacts from one location to another.
}
