% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ToolExitCode}
\alias{ToolExitCode}
\title{ToolExitCode Object}
\usage{
ToolExitCode(number = NULL)
}
\arguments{
\item{number}{Tool execution exit code}
}
\value{
ToolExitCode object
}
\description{
ToolExitCode Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Exit code from a tool execution.
}
