% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_objects.R
\name{ToolOutputReference}
\alias{ToolOutputReference}
\title{ToolOutputReference Object}
\usage{
ToolOutputReference(creationTime = NULL, output = NULL, testCase = NULL)
}
\arguments{
\item{creationTime}{The creation time of the file}

\item{output}{A FileReference to an output file}

\item{testCase}{The test case to which this output file belongs}
}
\value{
ToolOutputReference object
}
\description{
ToolOutputReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a ToolExecution output file.
}
