% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolresults_functions.R
\docType{package}
\name{toolresults_googleAuthR}
\alias{toolresults_googleAuthR}
\alias{toolresults_googleAuthR-package}
\title{Cloud Tool Results firstparty API
Reads and publishes results from Firebase Test Lab.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:20:00
filename: /Users/mark/dev/R/autoGoogleAPI/googletoolresultsv1beta3firstparty.auto/R/toolresults_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
